legacy/requirements.txt
liboqs-python==0.14.0
cryptography==45.0.6
